+++ packages/react-router-dom/server.tsx
@@ -6,6 +6,7 @@ import type {
   StaticHandlerContext,
   CreateStaticHandlerOptions as RouterCreateStaticHandlerOptions,
   UNSAFE_RouteManifest as RouteManifest,
+  RouterState,
 } from "@remix-run/router";
 import {
   IDLE_BLOCKER,
@@ -17,9 +18,16 @@ import {
   createStaticHandler as routerCreateStaticHandler,
   UNSAFE_convertRoutesToDataRoutes as convertRoutesToDataRoutes,
 } from "@remix-run/router";
-import { UNSAFE_mapRouteProperties as mapRouteProperties } from "react-router";
-import type { Location, RouteObject, To } from "react-router-dom";
-import { Routes } from "react-router-dom";
+import {
+  UNSAFE_mapRouteProperties as mapRouteProperties,
+  UNSAFE_useRoutesImpl as useRoutesImpl,
+} from "react-router";
+import type {
+  DataRouteObject,
+  Location,
+  RouteObject,
+  To,
+} from "react-router-dom";
 import {
   createPath,
   parsePath,
@@ -117,17 +125,20 @@ export function StaticRouterProvider({
     hydrateScript = `window.__staticRouterHydrationData = JSON.parse(${json});`;
   }
 
+  let { state } = dataRouterContext.router;
+
   return (
     <>
       <DataRouterContext.Provider value={dataRouterContext}>
-        <DataRouterStateContext.Provider value={dataRouterContext.router.state}>
+        <DataRouterStateContext.Provider value={state}>
           <Router
             basename={dataRouterContext.basename}
-            location={dataRouterContext.router.state.location}
-            navigationType={dataRouterContext.router.state.historyAction}
+            location={state.location}
+            navigationType={state.historyAction}
             navigator={dataRouterContext.navigator}
+            static={dataRouterContext.static}
           >
-            <Routes />
+            <DataRoutes routes={router.routes} state={state} />
           </Router>
         </DataRouterStateContext.Provider>
       </DataRouterContext.Provider>
@@ -140,6 +151,16 @@ export function StaticRouterProvider({
       ) : null}
     </>
   );
+}
+
+function DataRoutes({
+  routes,
+  state,
+}: {
+  routes: DataRouteObject[];
+  state: RouterState;
+}): React.ReactElement | null {
+  return useRoutesImpl(routes, undefined, state);
 }
 
 function serializeErrors(