+++ packages/router/router.ts
@@ -818,7 +818,7 @@ export function createRouter(init: RouterInit): Router {
 
   // Use this internal flag to force revalidation of all loaders:
   //  - submissions (completed or interrupted)
-  //  - useRevalidate()
+  //  - useRevalidator()
   //  - X-Remix-Revalidate (from redirect)
   let isRevalidationRequired = false;
 
@@ -3175,7 +3175,7 @@ function normalizeTo(
   let path = resolveTo(
     to ? to : ".",
     getPathContributingMatches(contextualMatches).map((m) => m.pathnameBase),
-    location.pathname,
+    stripBasename(location.pathname, basename) || location.pathname,
     relative === "path"
   );
 
@@ -3377,10 +3377,11 @@ function getMatchesToLoad(
       ...submission,
       actionResult,
       defaultShouldRevalidate:
-        // Forced revalidation due to submission, useRevalidate, or X-Remix-Revalidate
+        // Forced revalidation due to submission, useRevalidator, or X-Remix-Revalidate
         isRevalidationRequired ||
         // Clicked the same link, resubmitted a GET form
-        currentUrl.toString() === nextUrl.toString() ||
+        currentUrl.pathname + currentUrl.search ===
+          nextUrl.pathname + nextUrl.search ||
         // Search params affect all loaders
         currentUrl.search !== nextUrl.search ||
         isNewRouteInstance(currentRouteMatch, nextRouteMatch),
@@ -3436,7 +3437,7 @@ function getMatchesToLoad(
       nextParams: matches[matches.length - 1].params,
       ...submission,
       actionResult,
-      // Forced revalidation due to submission, useRevalidate, or X-Remix-Revalidate
+      // Forced revalidation due to submission, useRevalidator, or X-Remix-Revalidate
       defaultShouldRevalidate: isRevalidationRequired,
     });
     if (shouldRevalidate) {
@@ -4159,9 +4160,22 @@ function stripHashFromPath(path: To) {
 }
 
 function isHashChangeOnly(a: Location, b: Location): boolean {
-  return (
-    a.pathname === b.pathname && a.search === b.search && a.hash !== b.hash
-  );
+  if (a.pathname !== b.pathname || a.search !== b.search) {
+    return false;
+  }
+
+  if (a.hash === "") {
+    // No hash -> hash
+    return b.hash !== "";
+  } else if (a.hash === b.hash) {
+    // current hash -> same hash
+    return true;
+  } else if (b.hash !== "") {
+    // current hash -> new hash
+    return true;
+  }
+
+  return false;
 }
 
 function isDeferredResult(result: DataResult): result is DeferredResult {