+++ packages/react-router-dom-v5-compat/CHANGELOG.md
@@ -1,5 +1,21 @@
 # `react-router-dom-v5-compat`
 
+## 6.11.1
+
+### Patch Changes
+
+- Updated dependencies:
+  - `react-router@6.11.1`
+  - `react-router-dom@6.11.1`
+
+## 6.11.0
+
+### Patch Changes
+
+- Updated dependencies:
+  - `react-router@6.11.0`
+  - `react-router-dom@6.11.0`
+
 ## 6.10.0
 
 ### Patch Changes