+++ packages/react-router-native/CHANGELOG.md
@@ -1,5 +1,19 @@
 # `react-router-native`
 
+## 6.11.1
+
+### Patch Changes
+
+- Updated dependencies:
+  - `react-router@6.11.1`
+
+## 6.11.0
+
+### Patch Changes
+
+- Updated dependencies:
+  - `react-router@6.11.0`
+
 ## 6.10.0
 
 ### Patch Changes